ia-client/requirements.txt
keras==2.15.0
matplotlib==3.8.3
numpy==1.26.4
pandas==2.2.0
plotly==5.19.0
python-dotenv==1.0.1
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
sentry-sdk==1.41.0
st-pages==0.4.5
streamlit==1.31.1
tensorflow==2.15.0
